feat: lots of refactors
Move stuff around, ja know?
And fix routing on mobile.
And use the right widget for favorites.
And rename `BigListTile` to `FavListTile`.
And add a `main.dart` file that re-exports the production build.
And update file nesting again for the new naming scheme.
