+++ .vscode/settings.json
@@ -52,6 +52,6 @@
     ".replit": "replit.nix",
     "pubspec.yaml": ".metadata, analysis_options.yaml, pubspec.lock, build.yaml, dartdoc_options.yaml, l10n.yaml, flutter_launcher_icons*.yaml",
     "README.*": "*.md",
-    "*.dart": "${capture}.g.dart, ${capture}.gr.dart, ${capture}.freezed.dart"
+    "*.dart": "${capture}*.dart"
   }
 }
