+++ lib/main.dart
@@ -1,0 +1,1 @@
+export "main_production.dart";
