+++ lib/widgets/big_card/big_card.dart
