+++ lib/widgets/home/home.dart
@@ -2,10 +2,10 @@ import "package:auto_route/auto_route.dart";
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
-import "../model.dart";
+import "../../model.dart";
 
-import "../widgets/big_card.dart";
-import "../widgets/history.dart";
+import "../big_card/big_card.dart";
+import "../history/history.dart";
 
 @RoutePage()
 class GeneratorPage extends ConsumerWidget {
